Survival random forest with covariance estimation

      These trees will always be sampled without replacement.
                   The choices of tuning parameters will be limited.
                   Use at your own risk.

Usage

Surv_Cov_Forest(
  x,
  y,
  censor,
  testx,
  ntrees = 10000,
  mtry = max(1, as.integer(ncol(x)/3)),
  nmin = max(1, as.integer(log(nrow(x)))),
  split.gen = "best",
  nsplit = 0,
  resample.prob = 0.5,
  param.control = list(),
  ncores = 1,
  verbose = 0,
  seed = NULL,
  ...
)

Arguments

x

A matrix or data.frame of features

y

Response variable. a numeric/factor vector.

censor

The censoring indicator if survival model is used.

testx

A matrix or data.frame of testing data

ntrees

Number of trees. To obtain a stable and accurate estimation, the default is ntrees = 10000.

mtry

Number of randomly selected variables used at each internal node.

nmin

Terminal node size. Same as used in RLT.

split.gen

"random", "rank" or "best".

nsplit

Number of random cutting points to compare for each variable at an internal node.

resample.prob

Proportion of in-bag samples.

param.control

A list of additional parameters. However, choices are limited. For example, reinforcement splitting rules are not implemented.

ncores

Number of cores. Default is 0 (using all available cores).

verbose

Whether fitting info should be printed.

seed

Random seed number to replicate a previously fitted forest. Internally, the xoshiro256++ generator is used. If not specified, a seed will be generated automatically.

...

Additional arguments.

Value

           Prediction and covariance estimation of the testing data