check_param_RLT

Check parameters

The following parameters, if needed can be specified in the param.control list.

alpha can specify a minimum number of observations required for each child node as a portion of the parent node. Must be within [0, 0.5). When alpha $> 0$ and split.gen is rank or best, this will force each child node to contain at least \(\max(\texttt{nmin}, \alpha \times N_A)\) number of number of observations, where \(N_A\) is the sample size at the current internal node. This is mainly for theoretical concern.

split.rule specifies the splitting rule for comparisons. For regression, variance reduction "var" is used; for classification, "gini" index is used. For survival, "logrank", "suplogrank", and "coxgrad" are available. When "coxgrad" is used, variable weights "var.w" are used as the penalty.

VI.var=TRUE will calculate variance for variable importance. If VI.var=TRUE, resamp.prob(1-oob.prop) observations will be used for building each tree, resamp.proboob.prop observations will be used for the out-of-bag sample, and 1-resamp.prob observations will not be used as in-bag or out-of-bag samples for that tree at all. Only implemented for survival forests. Ignored otherwise.

oob.prop only used when VI.var=TRUE. Specifies the proportion of resamp.prob that will be assigned to the out-of-bag sample. Default is oob.prop=0.5. See VI.var for more details. Only implemented for survival forests. Ignored otherwise.

Usage

check_param_RLT(
  n,
  p,
  ntrees,
  mtry,
  nmin,
  split.gen,
  nsplit,
  resample.replace,
  resample.prob,
  resample.track,
  use.obs.w,
  use.var.w,
  linear.comb,
  importance,
  var.ready,
  ncores,
  verbose,
  reinforcement,
  param.control
)