get.surv.band 

Calculate the survival function (two-sided) confidence band from
             a RLT survival prediction.

Usage

get.surv.band(
  x,
  i = 0,
  alpha = 0.05,
  approach = "naive-mc",
  nsim = 1000,
  r = 3,
  ...
)

Arguments

x

A RLT prediction object. This must be an object calculated from a forest with var.ready = TRUE.

i

Observation number in the prediction. Default to calculate all (\(i = 0\))

alpha

alpha level for interval \((\alpha/2, 1 - \alpha/2)\)

approach

What approach is used to calculate the confidence band. Can be

naive-mc: positive-definite projection of the covariance matrix. the confidence band is non-smooth

smoothed-mc: use a smoothed marginal variance to perform the Monte Carlo approximation of the critical value. This is only recommended for large number of time points.

smoothed-lr: use a smoothed low-rank approximation of the covariance matrix and apply an adaptive Bonferroni correction to derive the critical values. Note that this approach relies on the assumption of the smoothness and low rank of the covariance matrix.

nsim

number of simulations for estimating the Monte Carlo critical value. Set this to be a large number. Default is 1000.

r

maximum number of ranks used in the smoothed-lr approximation. Usually 5 is enough for approximating the covariance matrix due to smoothness.

...

...