Reinforcement Learning Trees

      Fit models for regression, classification and 
                   survival analysis using reinforced splitting rules.
                   The model reduces to regular random forests if 
                   reinforcement is turned off.

If x is a data.frame, then all factors are treated as categorical variables.

To specify parameters of embedded models when reinforcement = TRUE, users can supply the following in the param.control list:

embed.resample.prob: proportion of samples (of the internal node) in the embedded model

embed.mtry: number or proportion of variables

embed.nmin terminal node size

embed.split.gen random cutting point search method ("random", "rank" or "best")

embed.nsplit number of random cutting points.

For some other experimental features, please see check_param_RLT.

Usage

RLT(
  x,
  y,
  censor = NULL,
  model = NULL,
  ntrees = if (reinforcement) 100 else 500,
  mtry = max(1, as.integer(ncol(x)/3)),
  nmin = max(1, as.integer(log(nrow(x)))),
  split.gen = "random",
  nsplit = 1,
  resample.replace = TRUE,
  resample.prob = if (resample.replace) 1 else 0.8,
  resample.preset = NULL,
  resample.track = FALSE,
  obs.w = NULL,
  var.w = NULL,
  linear.comb = 1,
  importance = FALSE,
  var.ready = FALSE,
  reinforcement = FALSE,
  param.control = list(),
  ncores = 0,
  verbose = 0,
  seed = NULL,
  ...
)

Arguments

x

A matrix or data.frame of features

y

Response variable. a numeric/factor vector.

censor

The censoring indicator if survival model is used.

model

The model type: "regression", "classification" or "survival".

ntrees

Number of trees, ntrees = 100 if reinforcement is used and ntrees = 1000 otherwise.

mtry

Number of randomly selected variables used at each internal node.

nmin

Terminal node size. Splitting will stop when the internal node size is less than twice of nmin. This is almost equivalent to setting nodesize \(= 2 \times\) nmin in the randomForest package.

split.gen

How the cutting points are generated: "random", "rank" or "best". If minimum child node size is enforced (alpha $> 0$), then "rank" and "best" should be used.

nsplit

Number of random cutting points to compare for each variable at an internal node.

resample.replace

Whether the in-bag samples are obtained with replacement.

resample.prob

Proportion of in-bag samples.

resample.preset

A pre-specified matrix for in-bag data indicator/count matrix. It must be an \(n \times\) ntrees matrix and cannot contain negative values. Extremely large counts are not recommended, and the sum of each column cannot exceed \(n\). If provided, then resample.track will be set to TRUE. This is an feature is mainly use when estimating variances of a random forest. Use at your own risk.

resample.track

Whether to keep track of the observations used in each tree.

obs.w

Observation weights

var.w

Variable weights. If this is supplied, the default is to perform weighted sampling of mtry variables. For other usage, see the details of split.rule in check_param_RLT.

linear.comb

When linear.comb is larger than 1, a linear combination split is used. When reinforcement is TRUE, the variables with the highest potential at an internal node is used. When reinforcement is FALSE, a marginal screening is used. In both cases, SIR and SAVE are used to determine the coefficients of the combination. When a categorical variable has the highest potential, then a single variable is used. Currently restricted to less than 5 number of variables in the linear combination.

importance

Whether to calculate variable importance measures. The calculation follows Breiman's original permutation strategy.

var.ready

Construct resample.preset automatically to allow variance estimations for prediction. If this is used, then resample.replace will be set to FALSE and resample.prob should be no larger than \(n / 2\). It is recommended to use a very large ntrees, e.g, 10000 or larger. For resample.prob greater than \(n / 2\), one should use the Reg_Var_Forest function.

reinforcement

Should reinforcement splitting rule be used. Default is "FALSE", i.e., regular random forests. When it is activated, embedded model tuning parameters are automatically chosen. They can also be specified in RLT.control.

param.control

A list of additional parameters. This can be used to specify other features in a random forest and set embedded model parameters for reinforcement splitting rules. See check_param_RLT and set_embed_param for more details. using reinforcement = TRUE will automatically generate some default tunings. However, they are not necessarily good.

ncores

Number of cores. Default is 0 (using all available cores).

verbose

Whether fitting info should be printed.

seed

Random seed number to replicate a previously fitted forest. Internally, the xoshiro256++ generator is used. If not specified, a seed will be generated automatically.

...

Additional arguments.

Value

A RLT object, constructed as a list consisting

FittedForest

Fitted tree structures

VarImp

Variable importance measures, if importance = TRUE

Prediction

In-bag prediction values

OOBPrediction

Out-of-bag prediction values

ObsTrack

An \(n \times\) ntrees matrix that indicates which observations are used in each tree. Provided if resample.preset is used or resample.track = TRUE.

For Survival Forests

NFail

The number of observed failure times

VarImpCov

if VI.var=TRUE, estimated covariance matrix for the variable importance

cindex_tree

Out-of-bag c-index for each tree

cindex

Out-of-bag c-index for the forest

timepoints

ordered observed failure times

y.point

order of \(y\) by observed failure times

References

Zhu, R., Zeng, D., & Kosorok, M. R. (2015) "Reinforcement Learning Trees." Journal of the American Statistical Association. 110(512), 1770-1784.