prediction using RLT

Predict the outcome (regression, classification or survival) using a fitted RLT object

Usage

# S3 method for class 'RLT'
predict(
  object,
  testx = NULL,
  var.est = FALSE,
  keep.all = FALSE,
  ncores = 1,
  verbose = 0,
  ...
)

Arguments

object

A fitted RLT object

testx

The testing samples, must have the same structure as the training samples

var.est

Whether to estimate the variance of each testing data. The original forest must be fitted with var.ready = TRUE. For survival forests, calculates the covariance matrix over all observed time points and calculates critical value for the confidence band.

keep.all

whether to keep the prediction from all trees. Warning: this can occupy a large storage space, especially in survival model

ncores

number of cores

verbose

print additional information

...

...

Value

A RLT prediction object, constructed as a list consisting

Prediction

Prediction

Variance

if var.est = TRUE and the fitted object is var.ready = TRUE

For Survival Forests

hazard

predicted hazard functions

CumHazard

predicted cumulative hazard function

Survival

predicted survival function

Allhazard

if keep.all = TRUE, the predicted hazard function for each observation and each tree

AllCHF

if keep.all = TRUE, the predicted cumulative hazard function for each observation and each tree

Cov

if var.est = TRUE and the fitted object is var.ready = TRUE. For each test subject, a matrix of size NFail\(\times\)NFail where NFail is the number of observed failure times in the training data

Var

if var.est = TRUE and the fitted object is var.ready = TRUE. Marginal variance for each subject

timepoints

ordered observed failure times from the training data

MarginalVar

if var.est = TRUE and the fitted object is var.ready = TRUE. Marginal variance for each subject from the Cov matrix projected to the nearest positive definite matrix

MarginalVarSmooth

if var.est = TRUE and the fitted object is var.ready = TRUE. Marginal variance for each subject from the Cov matrix projected to the nearest positive definite matrix and then smoothed using Gaussian kernel smoothing

CVproj

if var.est = TRUE and the fitted object is var.ready = TRUE. Critical values to calculate confidence bands around cumulative hazard predictions at several confidence levels. Calculated using MarginalVar

CVprojSmooth

if var.est = TRUE and the fitted object is var.ready = TRUE. Critical values to calculate confidence bands around cumulative hazard predictions at several confidence levels. Calculated using MarginalVarSmooth